Consolidate JDBCStoreResourceDefinition class hierarchy, since there are no longer multiple implementations.
